pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "facedetection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arize-phoenix-otel>=0.13.1",
    "fastmcp>=0.1.0",
    "ollama>=0.6.0",
    "openai>=2.3.0",
    "opencv-python>=4.11.0.86",
    "pytest>=8.4.2",
    "pytest-asyncio",
    "requests>=2.32.5",
    "retina-face>=0.0.17",
    "streamlit>=1.33.0",
    "watchdog>=6.0.0",
    "tensorflow-metal",
    "tf-keras>=2.16.0",
    "uvicorn>=0.37.0",
    "pydantic>=2.12.2",
    "mcp>=1.18.0",
    "google-adk>=1.16.0",
    "google-genai>=0.8.5",
]

[tool.setuptools.packages.find]
include = ["face_recognition*"]
